More safety with pointers using VLA function parameters

+++ tools/expected_distributions.h
@@ -167,7 +167,7 @@ check_table(uint64_t *exp, tableinfo_t *info)
 }
 
 static bool
-check_cocsep(uint64_t data_size, const void *data)
+check_cocsep(size_t data_size, const void *data)
 {
 	tableinfo_t info;
 
@@ -189,7 +189,7 @@ unknown_h48(uint8_t h, uint8_t k)
 }
 
 STATIC bool
-check_distribution(const char *solver, uint64_t data_size, const void *data)
+check_distribution(const char *solver, size_t data_size, const void *data)
 {
 	const char *str;
 	tableinfo_t info = {0};
